refactor: getTimeAgo 함수 utils함수로 분리
