+++ src/components/UserProfile.js
@@ -2,25 +2,9 @@ import React from "react";
 import styled from "styled-components";
 
 import SquareImage from "./SquareImage";
+import { getTimeAgo } from "../utils/utils";
 
 import DefaultProfileImage from "../assets/icon/profile.svg";
-
-function getTimeAgo(updatedAt) {
-  const currentTime = new Date();
-  const updateTime = new Date(updatedAt);
-
-  const differenceInMillis = currentTime - updateTime;
-  const differenceInHours = differenceInMillis / (1000 * 60 * 60);
-
-  if (differenceInHours < 1) {
-    return "방금 전";
-  } else if (differenceInHours < 24) {
-    return `${Math.floor(differenceInHours)}시간 전`;
-  } else {
-    const days = Math.floor(differenceInHours / 24);
-    return `${days}일 전`;
-  }
-}
 
 const UserProfile = ({ image, nickname = "똑똑한 판다", updatedAt }) => {
   const commentTimeAgo = getTimeAgo(updatedAt);
